Add files via upload
Dear Professor or Teach Assistants, 

I have done the assignment only I had the wrong kernel for it. So my code couldnt be processed. Please have a look at my code to see if it would have worked. I sent an email to the TA this (sunday) afternoon but have not gotten a response. I hope this will not affect my grade. 

Thanks in advance, 

Twan Duinmaijer